requirements.txt
# python==3.10.12
protobuf==5.26.1
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
torch==2.3.0
sentencepiece==0.2.0
transformers==4.32.1
tokenizers==0.13.2
datasets==2.12.0
accelerate==0.26.1
ensemble-boxes==1.0.8
PyYAML==6.0.1